backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic>=1.13.1
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
lightgbm>=4.5.0
joblib>=1.4.2
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
anthropic>=0.34.0
